# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted music streaming server API.",
        "title": "Korus API",
        "contact": {},
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/artists/{id}/image": {
            "get": {
                "description": "Returns artist image",
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Get artist image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist image",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/artwork/{id}": {
            "get": {
                "description": "Returns cover artwork image",
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "Streaming"
                ],
                "summary": "Get artwork for a track or album",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track or Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "track",
                            "album"
                        ],
                        "type": "string",
                        "default": "track",
                        "description": "Type of artwork",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artwork image",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/download/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download a track in original or transcoded format",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Streaming"
                ],
                "summary": "Download track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "mp3",
                            "aac",
                            "opus",
                            "flac",
                            "alac"
                        ],
                        "type": "string",
                        "description": "Target format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Bitrate in kbps",
                        "name": "bitrate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/lyrics/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns lyrics and synced lyrics if available",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Get lyrics for a track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/stream/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stream audio - serves original file or redirects to HLS manifest",
                "produces": [
                    "audio/*"
                ],
                "tags": [
                    "Streaming"
                ],
                "summary": "Stream a track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "aac",
                            "mp3",
                            "opus",
                            "flac",
                            "alac"
                        ],
                        "type": "string",
                        "description": "Target format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bitrate in kbps",
                        "name": "bitrate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Auth token for player",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio stream",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "307": {
                        "description": "Redirect to HLS manifest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}